BufferGeometry: Move computeTangents() into core. Use in GLTFLoader.
