+++ docs/examples/zh/utils/BufferGeometryUtils.html
@@ -16,14 +16,6 @@
 
 		<h2>Methods</h2>
 
-		<h3>[method:null computeTangents]( [param:BufferGeometry geometry] )</h3>
-		<p>
-		geometry -- A [page:BufferGeometry BufferGeometry] instance, which must have index, position, normal, and uv attributes.<br /><br />
-
-		Calculates and adds tangent attribute to a geometry.
-
-		</p>
-
 		<h3>[method:BufferGeometry mergeBufferGeometries]( [param:Array geometries], [param:Boolean useGroups] )</h3>
 		<p>
 		geometries -- Array of [page:BufferGeometry BufferGeometry] instances.<br />
